requirements.txt
flask
requests
grpcio
grpcio-tools
googleapis-common-protos
mcp
duckduckgo-search
ddgs
uvicorn
langchain
langchain-community
langgraph
langchain-google-genai
faiss-cpu
pypdf
httpx
streamlit
qrcode
pillow
psycopg2-binary==2.9.9
